Feat: Add Swagger to the project
