+++ docs/auth.yaml
@@ -1,0 +1,92 @@
+paths:
+ /auth/signup:
+   post:
+     tags:
+       - Auth
+     summary: User registration
+     description: Creates a new user, sends an email to the user and returns user ID and email.
+     produces:
+       - application/json
+     requestBody:
+       required: true
+       description: Provide required data to create a user account.
+       content:
+         application/json:
+           schema:
+             $ref: '#/components/signup'
+           example:
+             role: student
+             firstName: John
+             lastName: Doe
+             email: johndoe@gmail.com
+             password: qwer1234
+     responses:
+       201:
+         description: Created
+         content:
+           application/json:
+             schema:
+               $ref: '#/components/signupResponse'
+             example:
+               userId: 6255bc080a75adf9223df100
+               userEmail: johndoe@gmail.com
+       400:
+         description: Bad Request
+         content:
+           application/json:
+             schema:
+               $ref: '#/components/Error'
+             example:
+               status: 400
+               code: 'EMAIL_NOT_SENT'
+               message: 'Email has not been sent.'
+
+components:
+ signup:
+   type: object
+   properties:
+     role:
+       type: string
+       enum: [student, tutor]
+     firstName:
+       type: string
+     lastName:
+       type: string
+     email:
+       type: string
+     password:
+       type: string
+ login:
+   type: object
+   properties:
+     email:
+       type: string
+     password:
+       type: string
+ signupResponse:
+   type: object
+   properties:
+     userId:
+       type: string
+     userEmail:
+       type: string
+ accessToken:
+   type: object
+   properties:
+     accessToken:
+       type: string
+ forgotPassword:
+   type: object
+   properties:
+     email:
+       type: string
+ resetPassword:
+   type: object
+   properties:
+     password:
+       type: string
+ Error:
+   type: object
+   properties:
+     password:
+       type: string
