+++ src/initialization/initialization.js
@@ -1,32 +1,8 @@
 const express = require('express')
 const cors = require('cors')
 const cookieParser = require('cookie-parser')
-const swaggerJsDoc = require('swagger-jsdoc')
 const swaggerUI = require('swagger-ui-express')
-
-const swaggerOptions = {
-  definition: {
-    openapi: '3.0.0',
-    info: {
-      title: 'Express API for space2study',
-      version: '',
-      description: ''
-    },
-    servers: [
-      {
-        url: process.env.SERVER_URL
-      }
-    ],
-    components: {
-      securitySchemes: {
-        //cookieAuth or any other security schema
-      }
-    }
-  },
-  apis: ['./path to your API documentation file']
-}
-
-const swaggerSettings = swaggerJsDoc(swaggerOptions)
+const swaggerSettings = require('~/initialization/swaggerSettings')
 
 const {
   config: { CLIENT_URL }
@@ -36,6 +12,8 @@ const { createNotFoundError } = require('~/utils/errorsHelper')
 const errorMiddleware = require('~/middlewares/error')
 
 const initialization = (app) => {
+  app.use('/docs', swaggerUI.serve, swaggerUI.setup(swaggerSettings))
+
   app.use(express.json({ limit: '10mb' }))
   app.use(express.urlencoded({ extended: true }))
   app.use(cookieParser())
@@ -55,8 +33,6 @@ const initialization = (app) => {
   })
 
   app.use(errorMiddleware)
-
-  app.use('/route for your swagger docs', swaggerUI.serve, swaggerUI.setup(swaggerSettings))
 }
 
 module.exports = initialization
